# docs/dataforge-api-contract.yaml
openapi: 3.1.0
info:
  title: DataForge API
  version: 1.0.0
  description: >
    REST API for authenticated dataset upload, asynchronous processing,
    deterministic profiling, quality analysis, and bounded analytical queries.
    The MVP accepts CSV uploads only. XLSX remains a post-MVP format extension.
    Clients may send X-Request-ID; the API always returns X-Request-ID,
    propagates it to structured logs and applicable background-job messages,
    and generates one when the client does not provide it.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /datasets:
    parameters:
      - { $ref: '#/components/parameters/RequestId' }
    get:
      tags: [Datasets]
      operationId: listDatasets
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: Owned datasets.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DatasetPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Datasets]
      operationId: createDataset
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDatasetRequest' }
      responses:
        '201':
          description: Dataset created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dataset' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/DatasetNameConflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /datasets/{dataset_id}:
    parameters:
      - { $ref: '#/components/parameters/DatasetId' }
      - { $ref: '#/components/parameters/RequestId' }
    get:
      tags: [Datasets]
      operationId: getDataset
      responses:
        '200':
          description: Owned dataset.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dataset' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Datasets]
      operationId: deleteDataset
      responses:
        '204': { description: Dataset soft-deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
    DatasetId: { name: dataset_id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
...
    RequestId: { name: X-Request-ID, in: header, required: false, description: Client-supplied correlation ID. The API generates one when omitted and returns the effective value in the X-Request-ID response header., schema: { type: string, minLength: 1, maxLength: 128 } }
    Cursor: { name: cursor, in: query, required: false, schema: { type: string } }
    PageSize: { name: page_size, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
...
  responses:
    Unauthorized: { description: Authentication required or token invalid., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotFound: { description: Resource not found or not visible to this user., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    DatasetNameConflict:
      description: An active dataset with this name already exists for the authenticated owner. Do not retry automatically.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: conflict
              message: An active dataset with this name already exists.
              retryable: false
            request_id: req_01JEXAMPLE
    ValidationError: { description: Invalid request., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
    UUID: { type: string, format: uuid }
    Timestamp: { type: string, format: date-time }
...
    CreateDatasetRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        description: { type: [string, 'null'], maxLength: 2000 }
    Dataset:
      type: object
      required: [id, name, created_at, updated_at]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        name: { type: string }
        description: { type: [string, 'null'] }
        created_at: { $ref: '#/components/schemas/Timestamp' }
        updated_at: { $ref: '#/components/schemas/Timestamp' }
...
    PageInfo:
      type: object
      required: [has_more]
      properties:
        next_cursor: { type: [string, 'null'] }
        has_more: { type: boolean }
    DatasetPage:
      type: object
      required: [items, page_info]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Dataset' } }
        page_info: { $ref: '#/components/schemas/PageInfo' }
...
    Error:
      type: object
      required: [error, request_id]
      properties:
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code:
              type: string
              enum:
                - validation_error
                - unsupported_file_type
                - file_too_large
                - authentication_required
                - invalid_credentials
                - token_expired
                - forbidden
                - resource_not_found
                - conflict
                - invalid_state_transition
                - version_not_ready
                - query_rejected
                - rate_limited
                - storage_unavailable
                - database_unavailable
                - queue_unavailable
                - processing_failed_permanent
                - processing_retry_exhausted
                - internal_error
              example: validation_error
            message: { type: string }
            details: { type: object, additionalProperties: true }
            retryable: { type: boolean, description: Whether the client may retry according to the error model policy. }
        request_id: { type: string, description: Effective X-Request-ID value. }